// src/Interfaces/Command.ts
import Client from '../Client';
import { Message } from 'discord.js';

interface Run {
    (client: Client, message: Message, args: string[]);
}

export interface Command {
    name: string;
    description: string;
    category: string;
    aliases?: string[];
    visable?: boolean;
    run: Run;
}